requirements.txt
streamlit
phidata
groq
yfinance
duckduckgo-search
plotly
python-dotenv